Added Deck class, split card classes up
Started logic for deck class, which will be used
for nearly all objects in the game.

Started working on details for the various cards,
splitting different kinds of cards, and creating a
csv with all the Mythos cards in it, as well as a Python
file which will eventually take the csv and create the
skeletons for the c# file for those cards.